+++ src/modules/infra/handlers/graphql/mod.ts
@@ -1,0 +1,5 @@
+export { ZanixResolver } from './base.ts'
+export { Resolver } from './decorators/base.ts'
+export { Query } from './decorators/query.ts'
+export { Mutation } from './decorators/mutation.ts'
+export { Request } from './decorators/request.ts'